+++ changelog.md
@@ -5,6 +5,7 @@
 Features
 - Move the cleu logic to init.lua instead of exec.lua
 - Move pre- and post-decision logic to init.lua
+- Return objects instead of tuples for isAuraExpired and canCast
 
 #### v0.6 (2025-07-07)